octogenarian: refactor NoteRef as a Serpentine path
Replace the `NoteRef` case class and the `Divisible[GitHash, Text]`
bridge with `type NoteRef = Path on Notes under GitHash`.  A one-line
`extension (hash: GitHash) def /` at the package level triggers the
`Conversion[GitHash, CommitRoot]` from the previous commit and
immediately hands off to Serpentine's `Path.def /`; nothing in
Octogenarian does any path-construction work itself.

The previous `.target` and `.read[T]` members carry over as extension
methods on the path alias; `.ref` is renamed to `.namespace` and
materialises the canonical `refs/notes/<…>` path on demand.
