--- lib/octogenarian/src/core/octogenarian.NoteRef.scala
@@ -1,76 +1,0 @@
-                                                                                                  /*
-┏━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┓
-┃                                                                                                  ┃
-┃                                                   ╭───╮                                          ┃
-┃                                                   │   │                                          ┃
-┃                                                   │   │                                          ┃
-┃   ╭───────╮╭─────────╮╭───╮ ╭───╮╭───╮╌────╮╭────╌┤   │╭───╮╌────╮╭────────╮╭───────╮╭───────╮   ┃
-┃   │   ╭───╯│   ╭─╮   ││   │ │   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮   ││   ╭─╮  ││   ╭───╯│   ╭───╯   ┃
-┃   │   ╰───╮│   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╰─╯  ││   ╰───╮│   ╰───╮   ┃
-┃   ╰───╮   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   │ │   ││   ╭────╯╰───╮   │╰───╮   │   ┃
-┃   ╭───╯   ││   ╰─╯   ││   ╰─╯   ││   │ │   ││   ╰─╯   ││   │ │   ││   ╰────╮╭───╯   │╭───╯   │   ┃
-┃   ╰───────╯╰─────────╯╰────╌╰───╯╰───╯ ╰───╯╰────╌╰───╯╰───╯ ╰───╯╰────────╯╰───────╯╰───────╯   ┃
-┃                                                                                                  ┃
-┃    Soundness, version 0.54.0.                                                                    ┃
-┃    © Copyright 2021-25 Jon Pretty, Propensive OÜ.                                                ┃
-┃                                                                                                  ┃
-┃    The primary distribution site is:                                                             ┃
-┃                                                                                                  ┃
-┃        https://soundness.dev/                                                                    ┃
-┃                                                                                                  ┃
-┃    Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file     ┃
-┃    except in compliance with the License. You may obtain a copy of the License at                ┃
-┃                                                                                                  ┃
-┃        https://www.apache.org/licenses/LICENSE-2.0                                               ┃
-┃                                                                                                  ┃
-┃    Unless required by applicable law or agreed to in writing,  software distributed under the    ┃
-┃    License is distributed on an "AS IS" BASIS,  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,    ┃
-┃    either express or implied. See the License for the specific language governing permissions    ┃
-┃    and limitations under the License.                                                            ┃
-┃                                                                                                  ┃
-┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
-                                                                                                  */
-package octogenarian
-
-import ambience.*
-import anticipation.*
-import contingency.*
-import distillate.*
-import guillotine.*
-import prepositional.*
-import serpentine.*
-import symbolism.*
-
-import GitError.Reason.*
-
-// A `NoteRef` names a specific git note: the commit it is attached to, plus
-// the ref path (e.g. `refs/notes/ci-attestation`) it lives under. The repo
-// is supplied implicitly at `.read` time.
-//
-// Constructed via Symbolism's `/` operator: `commit / t"foo"` lifts a
-// `GitHash` into a `NoteRef` (given lives on `GitHash`'s companion in
-// `octogenarian.internal`), and further chaining (`noteRef / t"bar"`)
-// extends the namespace path via the given below.
-//
-// NB: this currently uses a `Divisible` instance for the entry-point lift
-// rather than a Serpentine-style `Conversion[GitHash, Path]`, because
-// Symbolism's generic `/` extension shadows any Conversion at method-
-// resolution time.  Revisit when Serpentine grows a first-class lift.
-object NoteRef:
-  given continuation: Tactic[GitRefError]
-  =>  NoteRef is Divisible by Text to NoteRef =
-
-    Divisible: (noteRef, segment) =>
-      GitRefs.validateSegment(segment)
-      NoteRef(noteRef.target, noteRef.ref / segment)
-
-case class NoteRef(target: GitHash, ref: Path on GitRefs):
-  def read[value: Decodable in Text]
-    ( using repo:    GitRepo,
-            command: GitCommand,
-            wd:      WorkingDirectory,
-            errors:  Tactic[GitError],
-            exec:    Tactic[ExecError] )
-  :   value logs GitEvent =
-
-    repo.notes.show(target, ref).lest(GitError(NoteNotFound)).decode[value]
